Update to support ebpf pkg v0.18.0 or higher
