+++ bpfprogs/probes.go
@@ -1,5 +1,8 @@
 // Copyright Contributors to the L3AF Project.
 // SPDX-License-Identifier: Apache-2.0
+//
+//go:build !WINDOWS
+// +build !WINDOWS
 
 package bpfprogs
